+++ package.json
@@ -27,8 +27,10 @@
     "config": "^1.30.0",
     "connect-session-sequelize": "^5.2.1",
     "cookie-parser": "^1.4.3",
+    "csvtojson": "^1.1.9",
     "diff": "^3.5.0",
     "elasticitems": "^1.0.16",
+    "elasticsearch": "^15.0.0",
     "eventemitter2": "^5.0.1",
     "express": "^4.16.2",
     "express-session": "^1.15.6",
@@ -38,6 +40,7 @@
     "json2csv": "^3.11.1",
     "lodash": "^4.17.5",
     "moment": "^2.21.0",
+    "multer": "^1.3.0",
     "nunjucks": "^3.1.2",
     "passport": "^0.4.0",
     "passport-google-oauth": "^1.0.0",
@@ -56,6 +59,7 @@
     "bower": "^1.8.2",
     "elasticbulk": "^1.0.7",
     "pg-native": "^2.2.0",
+    "supertest": "^3.1.0",
     "through2": "^2.0.3"
   }
 }
