+++ src/models/item.js
@@ -3,9 +3,15 @@ const Op = Sequelize.Op
 const sequelize = require('./../clients/sequelize');
 const User = require('./user');
 const _ = require('lodash');
+const config = require('config');
 const emitter = require('./../clients/emitter');
 
 const Item = sequelize.define('Item', {
+  id: {
+    type: Sequelize.INTEGER,
+    primaryKey: true,
+    autoIncrement: true,
+  },
   json: Sequelize.JSON,
   main_field: {
     type: Sequelize.STRING
@@ -22,6 +28,9 @@ const Item = sequelize.define('Item', {
   last_user_id: {
     type: Sequelize.INTEGER
   },
+  stamps: {
+    type: Sequelize.TEXT
+  },
   createdAt: Sequelize.DATE,
   updatedAt: Sequelize.DATE,
   deletedAt: Sequelize.DATE
@@ -31,14 +40,17 @@ const Item = sequelize.define('Item', {
 });
 
 Item.prototype.getItem = function() {
-  return _.merge(this.json, {
+  return _.merge(_.clone(this.json), {
     id: this.id
   })
 }
 
 Item.prototype.getElasticData = function() {
 
-  return this.json;
+  return _.merge(_.clone(this.json), {
+    _id: this.id,
+    id: this.id
+  })
   //return psqlToEsHelper.convertItem(this.dataValues);
 }
 
@@ -52,15 +64,38 @@ Item.findByIds = async function(ids) {
   });
 }
 
+Item.findByMainField = async function(main_field, options) {
+  if (!main_field) {
+    return null;
+  }
+
+  options = _.merge(options, {
+    where: {
+      main_field: main_field
+    }
+  })
+
+  return await Item.findOne(options);
+};
+
 Item.prototype.getExportData = function() {
 
-  return _.mapValues(this.json, function(o) {
+  var output = _.mapValues(this.json, function(o) {
     if (_.isArray(o)) {
       return o.join(',');
     }
 
     return o;
   });
+
+
+  var fields = config.get('export.fields');
+
+  if (fields) {
+    return _.pick(output, fields);
+  }
+
+  return output;
 }
 
 Item.prototype.getChangelogData = function(which) {
@@ -79,6 +114,22 @@ Item.prototype.getChangelogData = function(which) {
 /*Item.prototype.overrideJsonData = function(body) {
   Object.assign(this.json, body);
 }*/
+
+Item.hook('beforeUpdate', async (item, options) => {
+  var main_field = config.get('general.main_field');
+
+  if (main_field && item['json'][main_field]) {
+    item['main_field'] = item['json'][main_field];
+  }
+})
+
+Item.hook('beforeCreate', async (item, options) => {
+  var main_field = config.get('general.main_field');
+
+  if (main_field && item['json'][main_field]) {
+    item['main_field'] = item['json'][main_field];
+  }
+})
 
 Item.hook('afterUpdate', async (item, options) => {
   //console.log('after update');
