+++ test/fixtures/import_file_1.csv
@@ -1,0 +1,3 @@
+domain,tags,genre
+domain1.com,"a,b",Drama
+domain2.com,"a",Drama
