+++ test/fixtures/import_file_2.csv
@@ -1,0 +1,4 @@
+domain,tags
+domain1.com,"a,b,c"
+domain2.com,
+domain3.com,"d"
