+++ views/modals-content/item.html.twig
@@ -1,6 +1,7 @@
 <div class="modal-header">
   <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
   <h4 class="modal-title" id="myModalLabel">View item: {{ item.id }}</h4>
+  <span class="text-muted">Created: {{ row.createdAt | date('D MMM YYYY') }}, updated {{ row.updatedAt | date('D MMM YYYY') }}</span>
 </div>
 <div class="modal-body item-view">
   <div class="row" style="margin-bottom: 10px;">
