+++ views/modals-content/raw.html.twig
@@ -4,7 +4,7 @@
 </div>
 <div class="modal-body">
 
-  <a class="item-submit btn pull-right btn-default" onclick="showItem('{{ item["domain"] }}'); return false;">Back to view</a>
+  <a class="item-submit btn pull-right btn-default" onclick="showItem('{{ item["id"] }}'); return false;">Back to view</a>
   <div class="clearfix" style="margin-bottom: 10px;"></div>
 
   <h3>Raw data in PostgreSQL</h3>
